renamed RunningApp to RunningNode
